Move site adapters in submodule sites
