+++ tardis/adapter/sites/cloudstack.py
@@ -1,11 +1,11 @@
-from ..configuration.configuration import Configuration
-from ..exceptions.tardisexceptions import TardisTimeout
-from ..exceptions.tardisexceptions import TardisError
-from ..exceptions.tardisexceptions import TardisQuotaExceeded
-from ..exceptions.tardisexceptions import TardisResourceStatusUpdateFailed
-from ..interfaces.siteadapter import ResourceStatus
-from ..interfaces.siteadapter import SiteAdapter
-from ..utilities.staticmapping import StaticMapping
+from tardis.configuration.configuration import Configuration
+from tardis.exceptions.tardisexceptions import TardisTimeout
+from tardis.exceptions.tardisexceptions import TardisError
+from tardis.exceptions.tardisexceptions import TardisQuotaExceeded
+from tardis.exceptions.tardisexceptions import TardisResourceStatusUpdateFailed
+from tardis.interfaces.siteadapter import ResourceStatus
+from tardis.interfaces.siteadapter import SiteAdapter
+from tardis.utilities.staticmapping import StaticMapping
 
 from aiohttp import ClientConnectionError
 from cobald.daemon import runtime
