+++ tardis/adapter/sites/openstack.py
@@ -4,15 +4,15 @@ from simple_rest_client.exceptions import AuthError
 from simple_rest_client.exceptions import ClientError
 from aiohttp import ClientConnectionError
 from aiohttp import ContentTypeError
-from ..configuration.configuration import Configuration
-from ..exceptions.tardisexceptions import TardisAuthError
-from ..exceptions.tardisexceptions import TardisDroneCrashed
-from ..exceptions.tardisexceptions import TardisError
-from ..exceptions.tardisexceptions import TardisTimeout
-from ..exceptions.tardisexceptions import TardisResourceStatusUpdateFailed
-from ..interfaces.siteadapter import ResourceStatus
-from ..interfaces.siteadapter import SiteAdapter
-from ..utilities.staticmapping import StaticMapping
+from tardis.configuration.configuration import Configuration
+from tardis.exceptions.tardisexceptions import TardisAuthError
+from tardis.exceptions.tardisexceptions import TardisDroneCrashed
+from tardis.exceptions.tardisexceptions import TardisError
+from tardis.exceptions.tardisexceptions import TardisTimeout
+from tardis.exceptions.tardisexceptions import TardisResourceStatusUpdateFailed
+from tardis.interfaces.siteadapter import ResourceStatus
+from tardis.interfaces.siteadapter import SiteAdapter
+from tardis.utilities.staticmapping import StaticMapping
 
 from asyncio import TimeoutError
 from contextlib import contextmanager
